requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
altgraph==0.17.2
atomicwrites==1.4.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.1.1
future==0.18.2
idna==3.4
iniconfig==2.0.0
more-itertools==9.1.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pefile==2022.5.30
pluggy==0.13.1
py==1.11.0
pycryptodome==3.17
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
PyMySQL==1.0.2
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.7.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
wcwidth==0.2.6
